+++ packages/sdk38/src/lcdapi/staking.spec.ts
@@ -2,7 +2,7 @@
 import { assert, sleep } from "@cosmjs/utils";
 
 import { coin, coins } from "../coins";
-import { isPostTxFailure } from "../cosmosclient";
+import { assertIsPostTxSuccess } from "../cosmosclient";
 import { makeSignBytes } from "../encoding";
 import { MsgDelegate, MsgUndelegate } from "../msgs";
 import { SigningCosmosClient } from "../signingcosmosclient";
@@ -56,8 +56,8 @@ describe("StakingExtension", () => {
           signatures: [signature],
         };
 
-        const receipt = await client.postTx(tx);
-        assert(!isPostTxFailure(receipt));
+        const result = await client.postTx(tx);
+        assertIsPostTxSuccess(result);
       }
       {
         const msg: MsgUndelegate = {
@@ -79,8 +79,8 @@ describe("StakingExtension", () => {
           signatures: [signature],
         };
 
-        const receipt = await client.postTx(tx);
-        assert(!isPostTxFailure(receipt));
+        const result = await client.postTx(tx);
+        assertIsPostTxSuccess(result);
       }
 
       await sleep(75); // wait until transactions are indexed
@@ -402,7 +402,7 @@ describe("StakingExtension", () => {
       const response = await client.staking.validatorDelegations(validatorAddress);
       expect(response).toEqual({
         height: jasmine.stringMatching(nonNegativeIntegerMatcher),
-        result: [
+        result: jasmine.arrayContaining([
           {
             delegator_address: faucet.address,
             validator_address: validatorAddress,
@@ -415,7 +415,7 @@ describe("StakingExtension", () => {
             shares: "250000000.000000000000000000",
             balance: { denom: "ustake", amount: "250000000" },
           },
-        ],
+        ]),
       });
     });
   });