+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - @cosmjs/cosmwasm: Rename `CosmWasmClient.getNonce` method to `.getSequence`.
 - @cosmjs/cosmwasm: Remove `RestClient` class in favour of new modular
   `LcdClient` class from @cosmjs/sdk38.
+- @cosmjs/cosmwasm: Add `SigningCosmWasmClient.signAndPost` as a mid-level
+  abstraction between `SigningCosmWasmClient.upload`/`.instantiate`/`.execute`
+  and `.postTx`.
+- @cosmjs/cosmwasm: Use `*PostTx*` types and helpers from @cosmjs/sdk38. Remove
+  exported `PostTxResult`.
 - @cosmjs/sdk38: Rename `CosmosClient.getNonce` method to `.getSequence`.
 - @cosmjs/sdk38: Remove `RestClient` class in favour of new modular `LcdClient`
   class.
@@ -13,3 +18,7 @@
   `OfflineSigner` instead of a `SigningCallback`.
 - @cosmjs/math: Add missing integer check to `Uint64.fromNumber`. Before
   `Uint64.fromNumber(1.1)` produced some result.
+- @cosmjs/sdk38: Add `SigningCosmosClient.signAndPost` as a mid-level
+  abstraction between `SigningCosmosClient.sendTokens` and `.postTx`.
+- @cosmjs/sdk38: Export `PostTxFailure`/`PostTxSuccess` and type checkers
+  `isPostTxFailure`/`isPostTxSuccess`; export `assertIsPostTxSuccess`.